# Conflicts:
#	react-client/hogar-client/package-lock.json
#	react-client/hogar-client/package.json
#	react-client/hogar-client/src/components/EmployeeCard.tsx
#	react-client/hogar-client/src/views/ContactUs.tsx
#	react-client/hogar-client/src/views/Home.tsx

+++ paw-webapp/model/src/main/java/ar/edu/itba/paw/model/Employee.java
@@ -92,7 +92,7 @@ public class Employee implements Serializable {
         return rating;
     }
 
-    public void setRating(long rating) {
+    public void setRating(float rating) {
         this.rating = rating;
     }
